+++ gatsby-config.js
@@ -8,7 +8,7 @@ module.exports = {
     {
       resolve: "gatsby-plugin-transition-link",
       options: {
-        layout: require.resolve(`./src/components/Layout.js`),
+        layout: require.resolve(`./src/components/layout/index.js`),
       },
     },
     `gatsby-plugin-sass`,
